refactor executor to remove dependency on hardhat
