+++ packages/core/package.json
@@ -34,6 +34,7 @@
     "ethers": "^5.6.9",
     "handlebars": "^4.7.7",
     "hardhat-deploy": "^0.11.18",
+    "ipfs-http-client": "56.0.3",
     "merkletreejs": "^0.2.18"
   }
 }
