+++ packages/core/src/actions/index.ts
@@ -1,2 +1,3 @@
 export * from './bundle'
 export * from './types'
+export * from './artifacts'
