+++ packages/core/src/config/fetch.ts
@@ -1,0 +1,47 @@
+import { create, IPFSHTTPClient } from 'ipfs-http-client'
+
+import { CanonicalChugSplashConfig } from './types'
+
+export const chugsplashFetchSubtask = async (args: {
+  configUri: string
+  ipfsUrl?: string
+}): Promise<CanonicalChugSplashConfig> => {
+  let config: CanonicalChugSplashConfig
+  let ipfs: IPFSHTTPClient
+  if (args.ipfsUrl) {
+    ipfs = create({
+      url: args.ipfsUrl,
+    })
+  } else if (process.env.IPFS_PROJECT_ID && process.env.IPFS_API_KEY_SECRET) {
+    const projectCredentials = `${process.env.IPFS_PROJECT_ID}:${process.env.IPFS_API_KEY_SECRET}`
+    ipfs = create({
+      host: 'ipfs.infura.io',
+      port: 5001,
+      protocol: 'https',
+      headers: {
+        authorization: `Basic ${Buffer.from(projectCredentials).toString(
+          'base64'
+        )}`,
+      },
+    })
+  } else {
+    throw new Error(
+      'You must either set your IPFS credentials in an environment file or call this task with an IPFS url.'
+    )
+  }
+
+  if (args.configUri.startsWith('ipfs://')) {
+    const decoder = new TextDecoder()
+    let data = ''
+    const stream = await ipfs.cat(args.configUri.replace('ipfs://', ''))
+    for await (const chunk of stream) {
+      // Chunks of data are returned as a Uint8Array. Convert it back to a string
+      data += decoder.decode(chunk, { stream: true })
+    }
+    config = JSON.parse(data)
+  } else {
+    throw new Error('unsupported URI type')
+  }
+
+  return config
+}
