+++ packages/core/src/config/index.ts
@@ -1,2 +1,3 @@
 export * from './parse'
 export * from './types'
+export * from './fetch'
