+++ packages/plugins/src/hardhat/artifacts.ts
@@ -4,19 +4,11 @@ import * as semver from 'semver'
 import {
   SolidityStorageLayout,
   ChugSplashConfig,
-  CanonicalChugSplashConfig,
   createDeploymentFolderForNetwork,
   writeDeploymentArtifact,
+  getConstructorArgs,
 } from '@chugsplash/core'
-import { add0x, remove0x } from '@eth-optimism/core-utils'
-import { ethers, utils } from 'ethers'
-import { SolcBuild } from 'hardhat/types'
-import { getCompilersDir } from 'hardhat/internal/util/global-dir'
-import {
-  CompilerDownloader,
-  CompilerPlatform,
-} from 'hardhat/internal/solidity/compiler/downloader'
-import { Compiler, NativeCompiler } from 'hardhat/internal/solidity/compiler'
+import { ethers } from 'ethers'
 
 // TODO
 export type ContractArtifact = any
@@ -103,326 +95,12 @@ https://github.com/ethereum-optimism/smock#note-on-using-smoddit`
   return (output as any).storageLayout
 }
 
-export const getCreationCode = (
-  bytecode: string,
-  parsedConfig: ChugSplashConfig,
-  referenceName: string,
-  abi: any,
-  compilerOutput: any,
-  sourceName: string,
-  contractName: string
-): string => {
-  const { constructorArgTypes, constructorArgValues } = getConstructorArgs(
-    parsedConfig,
-    referenceName,
-    abi,
-    compilerOutput,
-    sourceName,
-    contractName
-  )
-
-  const creationCodeWithConstructorArgs = bytecode.concat(
-    remove0x(
-      utils.defaultAbiCoder.encode(constructorArgTypes, constructorArgValues)
-    )
-  )
-
-  return creationCodeWithConstructorArgs
-}
-
-// TODO: I think this should go in /core now that we don't rely on hardhat as a dependency.
-// We could potentially move other contracts in here to core too.
-export const getConstructorArgs = (
-  parsedConfig: ChugSplashConfig,
-  referenceName: string,
-  abi: any,
-  compilerOutput: any,
-  sourceName: string,
-  contractName: string
-): { constructorArgTypes: any[]; constructorArgValues: any[] } => {
-  const immutableReferences =
-    compilerOutput.contracts[sourceName][contractName].evm.deployedBytecode
-      .immutableReferences
-
-  const contractConfig = parsedConfig.contracts[referenceName]
-
-  const constructorFragment = abi.find(
-    (fragment) => fragment.type === 'constructor'
-  )
-  const constructorArgTypes = []
-  const constructorArgValues = []
-  if (constructorFragment === undefined) {
-    return { constructorArgTypes, constructorArgValues }
-  }
-
-  // Maps a constructor argument name to the corresponding variable name in the ChugSplash config
-  const constructorArgNamesToImmutableNames = {}
-
-  for (const source of Object.values(compilerOutput.sources)) {
-    for (const contractNode of (source as any).ast.nodes) {
-      if (
-        contractNode.nodeType === 'ContractDefinition' &&
-        contractNode.nodes !== undefined
-      ) {
-        for (const node of contractNode.nodes) {
-          if (
-            node.nodeType === 'VariableDeclaration' &&
-            node.mutability === 'immutable' &&
-            Object.keys(immutableReferences).includes(node.id.toString(10))
-          ) {
-            if (contractConfig.variables[node.name] === undefined) {
-              throw new Error(
-                `Could not find immutable variable "${node.name}" in ${referenceName}.
-Did you forget to declare it in ${parsedConfig.options.projectName}?`
-              )
-            }
-
-            const constructorArgName =
-              getConstructorArgNameForImmutableVariable(
-                contractConfig.contract,
-                contractNode.nodes,
-                node.name
-              )
-            constructorArgNamesToImmutableNames[constructorArgName] = node.name
-          }
-        }
-      }
-    }
-  }
-
-  constructorFragment.inputs.forEach((input) => {
-    constructorArgTypes.push(input.type)
-    if (constructorArgNamesToImmutableNames.hasOwnProperty(input.name)) {
-      constructorArgValues.push(
-        contractConfig.variables[
-          constructorArgNamesToImmutableNames[input.name]
-        ]
-      )
-    } else {
-      throw new Error(
-        `Detected a non-immutable constructor argument, "${input.name}", in ${contractConfig.contract}.
-Please remove it or make the corresponding variable immutable.`
-      )
-    }
-  })
-
-  return { constructorArgTypes, constructorArgValues }
-}
-
 export const getDeployedBytecode = async (
   provider: ethers.providers.JsonRpcProvider,
   address: string
 ): Promise<string> => {
   const deployedBytecode = await provider.getCode(address)
   return deployedBytecode
-}
-
-export const getNestedConstructorArg = (variableName: string, args): string => {
-  let remainingArguments = args[0]
-  while (remainingArguments !== undefined) {
-    if (remainingArguments.name !== undefined) {
-      return remainingArguments.name
-    }
-    remainingArguments = remainingArguments.arguments[0]
-  }
-  throw new Error(
-    `Could not find nested constructor argument for the immutable variable ${variableName}.
-Please report this error.`
-  )
-}
-
-export const getConstructorArgNameForImmutableVariable = (
-  contractName: string,
-  nodes: any,
-  variableName: string
-): string => {
-  for (const node of nodes) {
-    if (node.kind === 'constructor') {
-      for (const statement of node.body.statements) {
-        if (statement.expression.nodeType === 'FunctionCall') {
-          throw new Error(
-            `Please remove the "${statement.expression.expression.name}" call in the constructor for ${contractName}.`
-          )
-        } else if (statement.expression.nodeType !== 'Assignment') {
-          throw new Error(
-            `disallowed statement in constructor for ${contractName}: ${statement.expression.nodeType}`
-          )
-        }
-        if (statement.expression.leftHandSide.name === variableName) {
-          if (typeof statement.expression.rightHandSide.name === 'string') {
-            return statement.expression.rightHandSide.name
-          } else if (
-            statement.expression.rightHandSide.kind === 'typeConversion'
-          ) {
-            return getNestedConstructorArg(
-              variableName,
-              statement.expression.rightHandSide.arguments
-            )
-          } else {
-            throw new Error(
-              `The immutable variable "${variableName}" must be assigned directly to a
-constructor argument inside the body of the constructor in ${contractName}.`
-            )
-          }
-        }
-      }
-    }
-  }
-  throw new Error(
-    `Could not find immutable variable assignment for ${variableName}.
-Did you forget to include it in your ChugSplash config file?`
-  )
-}
-
-export const getImmutableVariables = (
-  compilerOutput: any,
-  sourceName: string,
-  contractName: string
-): string[] => {
-  const immutableReferences: {
-    [astId: number]: {
-      length: number
-      start: number
-    }[]
-  } =
-    compilerOutput.contracts[sourceName][contractName].evm.deployedBytecode
-      .immutableReferences
-
-  if (
-    immutableReferences === undefined ||
-    Object.keys(immutableReferences).length === 0
-  ) {
-    return []
-  }
-
-  const immutableVariables: string[] = []
-  for (const source of Object.values(compilerOutput.sources)) {
-    for (const contractNode of (source as any).ast.nodes) {
-      if (contractNode.nodeType === 'ContractDefinition') {
-        for (const node of contractNode.nodes) {
-          if (node.mutability === 'immutable') {
-            immutableVariables.push(node.name)
-          }
-        }
-      }
-    }
-  }
-  return immutableVariables
-}
-
-// Credit: NomicFoundation
-// https://github.com/NomicFoundation/hardhat/blob/main/packages/hardhat-core/src/builtin-tasks/compile.ts
-export const getSolcBuild = async (solcVersion: string) => {
-  const compilersCache = await getCompilersDir()
-
-  const compilerPlatform = CompilerDownloader.getCompilerPlatform()
-  const downloader = CompilerDownloader.getConcurrencySafeDownloader(
-    compilerPlatform,
-    compilersCache
-  )
-
-  const isCompilerDownloaded = await downloader.isCompilerDownloaded(
-    solcVersion
-  )
-
-  if (!isCompilerDownloaded) {
-    console.log(`Downloading compiler version ${solcVersion}`)
-    await downloader.downloadCompiler(solcVersion)
-  }
-
-  const compiler = await downloader.getCompiler(solcVersion)
-
-  if (compiler !== undefined) {
-    return compiler
-  }
-
-  const wasmDownloader = CompilerDownloader.getConcurrencySafeDownloader(
-    CompilerPlatform.WASM,
-    compilersCache
-  )
-
-  const isWasmCompilerDownloader = await wasmDownloader.isCompilerDownloaded(
-    solcVersion
-  )
-
-  if (!isWasmCompilerDownloader) {
-    console.log(`Downloading compiler version ${solcVersion}`)
-    await wasmDownloader.downloadCompiler(solcVersion)
-  }
-
-  const wasmCompiler = await wasmDownloader.getCompiler(solcVersion)
-  return wasmCompiler
-}
-
-export const getArtifactsFromParsedCanonicalConfig = async (
-  parsedCanonicalConfig: CanonicalChugSplashConfig
-): Promise<{ [referenceName: string]: any }> => {
-  const compilerOutputs: any[] = []
-  // Get the compiler output for each compiler input.
-  for (const compilerInput of parsedCanonicalConfig.inputs) {
-    const solcBuild: SolcBuild = await getSolcBuild(compilerInput.solcVersion)
-
-    let compilerOutput: any // TODO: Compiler output type
-    if (solcBuild.isSolcJs) {
-      const compiler = new Compiler(solcBuild.compilerPath)
-      compilerOutput = await compiler.compile(compilerInput.input)
-    } else {
-      const compiler = new NativeCompiler(solcBuild.compilerPath)
-      compilerOutput = await compiler.compile(compilerInput.input)
-    }
-    compilerOutputs.push(compilerOutput)
-  }
-
-  const artifacts = {}
-  // Generate an artifact for each contract in the ChugSplash config.
-  for (const [referenceName, contractConfig] of Object.entries(
-    parsedCanonicalConfig.contracts
-  )) {
-    let compilerOutputIndex = 0
-    while (artifacts[referenceName] === undefined) {
-      // Iterate through the sources in the current compiler output to find the one that
-      // contains this contract.
-      const compilerOutput = compilerOutputs[compilerOutputIndex]
-      for (const [sourceName, sourceOutput] of Object.entries(
-        compilerOutput.contracts
-      )) {
-        // Check if the current source contains the contract.
-        if (sourceOutput.hasOwnProperty(contractConfig.contract)) {
-          const contractOutput = sourceOutput[contractConfig.contract]
-
-          const creationCode = getCreationCode(
-            add0x(contractOutput.evm.bytecode.object),
-            parsedCanonicalConfig,
-            referenceName,
-            contractOutput.abi,
-            compilerOutput,
-            sourceName,
-            contractConfig.contract
-          )
-          const immutableVariables = getImmutableVariables(
-            compilerOutput,
-            sourceName,
-            contractConfig.contract
-          )
-
-          artifacts[referenceName] = {
-            creationCode,
-            storageLayout: contractOutput.storageLayout,
-            immutableVariables,
-            abi: contractOutput.abi,
-            compilerOutput,
-            sourceName,
-            contractName: contractConfig.contract,
-          }
-          // We can exit the loop at this point since each contract only has a single artifact
-          // associated with it.
-          break
-        }
-      }
-      compilerOutputIndex += 1
-    }
-  }
-  return artifacts
 }
 
 export const createDeploymentArtifacts = async (
